requirements.txt
./vindinium-python
keras-rl
tensorflow
gym